requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
exceptiongroup==1.2.1
filelock==3.15.4
h11==0.14.0
identify==2.5.36
idna==3.7
lxml==5.2.2
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==24.1
platformdirs==4.2.2
pre-commit==3.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
